% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_regression.R
\name{check_sibling_order}
\alias{check_sibling_order}
\title{Check Sibling Order}
\usage{
check_sibling_order(..., fast = FALSE)
}
\arguments{
\item{...}{Additional arguments to be passed to the function.}

\item{fast}{Logical. If TRUE, uses a faster method for data processing.}
}
\value{
A one-row data frame with a new column order indicating which familial member (1, 2, or
  neither) has more of the outcome.
}
\description{
This function determines the order of sibling pairs based on an outcome variable.
The function checks which of the two kinship pairs has more of a specified outcome variable.
It adds a new column named `order` to the dataset, indicating which sibling (identified as "s1" or "s2") has more of the outcome.
If the two siblings have the same amount of the outcome, it randomly assigns one as having more.
}
