% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizing.R
\name{checkSiblingOrderUpdating}
\alias{checkSiblingOrderUpdating}
\title{Check which sibling has more of the outcome}
\usage{
checkSiblingOrderUpdating(data, outcome, row)
}
\arguments{
\item{data}{The output of \link[Nlsylinks]{CreatePairLinksDoubleEntered}.}

\item{outcome}{A character string containing the outcome variable of
interest.}

\item{row}{The row number of the data frame}
}
\value{
A character string signifying which familial member (1, 2, or
  neither) has more of the outcome.
}
\description{
This function takes in the output of
\link[Nlsylinks]{CreatePairLinksDoubleEntered} and adds a column \code{order}
by comparing which familial member has more of the outcome. This is done per
pair (i.e. row).
}
