% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_flu_ses}
\alias{data_flu_ses}
\title{Flu Vaccination and SES Data}
\format{
A data frame. \describe{Kinship pairs and their relatedness,
SES, and flu vaccination information.}
}
\source{
NLSY/R Lab
}
\usage{
data_flu_ses
}
\description{
A data frame that accompanies the regression vignette.
It contains data on SES and flu vaccination.
}
\keyword{datasets}
