% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_Functions.R
\name{detachAllData}
\alias{detachAllData}
\title{Collection of Miscellaneous functions.}
\usage{
detachAllData()
}
\description{
The following functions are ones I have found useful. Many of these (but not all) these functions are adapted from elsewhere. I have tried to credit the function with the source.
}

