% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_discord_data.R
\name{discord_data_ram_optimized}
\alias{discord_data_ram_optimized}
\title{Discord Data RAM Optimized}
\usage{
discord_data_ram_optimized(
  data,
  outcome,
  predictors,
  id = NULL,
  sex = "sex",
  race = "race",
  pair_identifiers,
  demographics = "both",
  coding_method = "none"
)
}
\description{
This function restructures data to determine kinship differences.
}
