% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord_power.R
\name{discord_power}
\alias{discord_power}
\title{Power Analysis}
\usage{
discord_power(
  r_all = c(1, 0.5),
  npg_all = 500,
  npergroup_all = rep(npg_all, length(r_all)),
  pheno_correlation = 0.5,
  variables = 2,
  r_vector = NULL,
  ace_all = c(1, 1, 1),
  ace_list = matrix(rep(ace_all, variables), byrow = TRUE, nrow = variables),
  cov_a = 0,
  cov_c = 0,
  cov_e = 0,
  alpha = 0.05,
  ...
)
}
\value{
Returns \code{data.frame} with the following:
}
\description{
Conduct Power Analysis for discord model
}
