% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{discord_regression_legacy}
\alias{discord_regression_legacy}
\title{Legacy Code: Discord Regression}
\usage{
discord_regression_legacy(
  df,
  outcome,
  predictors,
  more_args = NULL,
  additional_formula = more_args,
  ...
)
}
\arguments{
\item{outcome}{A character string containing the outcome variable of
interest.}

\item{predictors}{A character vector containing the column names for
predicting the outcome.}

\item{more_args}{Optional string to add additional inputs to formula}

\item{additional_formula}{Deprecated}
}
\value{
Resulting `lm` object from performing the discordant regression.
}
\description{
This is from
\url{https://github.com/R-Computing-Lab/discord/blob/74323b2cdd739355cd4a388251c747f1bcd87eb5/R/discord_regression.R}
and is used to perform the discordant regression on the data output from
\code{\link{discord_data_legacy}}.
}
\keyword{internal}
