% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord_data_JT.R
\name{discordData}
\alias{discordData}
\title{Prepare data for regression analysis}
\usage{
discordData(
  data,
  outcome,
  predictors,
  id = "extended_id",
  sex = "sex",
  race = "race"
)
}
\arguments{
\item{data}{The output of \code{\link{checkSiblingOrder}}.}

\item{outcome}{A character string containing the outcome variable of
interest.}

\item{predictors}{A character vector containing the column names for
predicting the outcome.}

\item{id}{The extended family pair ID from
\link[Nlsylinks]{CreatePairLinksDoubleEntered}. Should be a character
string.}

\item{sex}{The character string for the sex column name.}

\item{race}{The character string for the race column name.}
}
\value{
A data frame that contains one row for each family pair and the difference
and mean of their scores for different variables.
}
\description{
This function takes in the data, outcome, and predictors, and transforms it
into a format where a discord regression can be performed.
}
\examples{

# Get unique ExtendedIDs and join with the sampleData dataframe.
uniqueExtendedIDs <- sampleData \%>\%
dplyr::count(extended_id) \%>\%
dplyr::filter(n == 1) \%>\%
dplyr::select(-n) \%>\%
dplyr::left_join(sampleData)

discord:::discordData(data = uniqueExtendedIDs, outcome = "flu_2008",
predictors = c("edu_2008", "tnfi_2008"),
id = "extended_id", sex = "sex", race = "race")

}
