% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord_data_JT.R
\name{discordRegression}
\alias{discordRegression}
\title{Perform a Linear Regression within the Discordant Kinship Framework}
\usage{
discordRegression(preppedData, outcome, predictors)
}
\arguments{
\item{preppedData}{The output of \code{\link{discordData}}.}

\item{outcome}{A character string containing the outcome variable of
interest.}

\item{predictors}{A character vector containing the column names for
predicting the outcome.}
}
\value{
A tidy dataframe containing the model metrics via the
  \link[broom]{tidy} function.
}
\description{
Perform a Linear Regression within the Discordant Kinship Framework
}
\examples{

uniqueExtendedIDs <- sampleData \%>\%
dplyr::count(ExtendedID) \%>\%
dplyr::filter(n == 1) \%>\%
dplyr::select(-n) \%>\%
dplyr::left_join(sampleData)
test <- discordData(data = uniqueExtendedIDs,
variables = c("FLU_2008", "edu_2008", "tnfi_2008"),
id = "ExtendedID", sex = "SEX", race = "RACE")
fitModel <-
discordRegression(preppedData = test, outcome = "FLU_2008",
predictors = c("edu_2008", "tnfi_2008"))
}
