% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizing.R
\name{discordRegressionUpdating}
\alias{discordRegressionUpdating}
\title{Perform a Linear Regression within the Discordant Kinship Framework - Optimized}
\usage{
discordRegressionUpdating(
  data,
  outcome,
  predictors,
  id = "extended_id",
  sex = "sex",
  race = "race",
  pair_identifiers = c("_s1", "_s2")
)
}
\arguments{
\item{data}{The output of
\link[Nlsylinks]{CreatePairLinksSingleEntered}.}

\item{outcome}{A character string containing the outcome variable of
interest.}

\item{predictors}{A character vector containing the column names for
predicting the outcome.}

\item{id}{The extended family pair ID from
\link[Nlsylinks]{CreatePairLinksDoubleEntered}. Should be a character
string.}

\item{sex}{The character string for the sex column name.}

\item{race}{The character string for the race column name.}

\item{pair_identifiers}{A character vector of length two that contains the variable identifier for each kinship pair.}
}
\value{
A tidy dataframe containing the model metrics via the
  \link[broom]{tidy} function.
}
\description{
Perform a Linear Regression within the Discordant Kinship Framework - Optimized
}
\examples{

uniqueExtendedIDs <- sampleData \%>\%
dplyr::count(extended_id) \%>\%
dplyr::filter(n == 1) \%>\%
dplyr::select(-n) \%>\%
dplyr::left_join(sampleData)

fitModel <- discordRegressionUpdating(data = uniqueExtendedIDs, outcome = "flu_2008",
predictors = c("edu_2008", "tnfi_2008"),
pair_identifiers = c("_s1", "_s2"))
}
