% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord_data_JT.R
\name{makeMeanDiffs}
\alias{makeMeanDiffs}
\title{Add a difference and mean column}
\usage{
makeMeanDiffs(data, id, sex, race, variable, row)
}
\arguments{
\item{data}{The output of \code{\link{checkSiblingOrder}}.}

\item{id}{The extended family pair ID from
\link[Nlsylinks]{CreatePairLinksDoubleEntered}. Should be a character
vector.}

\item{sex}{The character string for the sex column name.}

\item{race}{The character string for the race column name.}

\item{variable}{The variable to get the mean and difference for.}

\item{row}{The row number of the data frame.}
}
\value{
This function returns a dataframe with four columns and one row. The
  first two columns are the variable with a "_1" or "_2". The former refers
  to the family member with the higher amount of the variable (first member),
  relative to the latter (second member). The variable's difference and mean
  for the family pair are included in the other two columns, in addition to
  the pairs' sex and race.
}
\description{
This function takes in a dataframe from the \code{\link{checkSiblingOrder}}
function. It also takes in a variable name and row number. It computes the
difference between the two family members for a given variable and their
mean.
}
