% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moreUniqueExtendedIDs}
\alias{moreUniqueExtendedIDs}
\title{More Unique IDs Derived from sampleData}
\format{
A dataframe for testing with 1200 rows (just multiples of uniqueExtendedIDs)
}
\source{
NLSY/R Lab
}
\usage{
moreUniqueExtendedIDs
}
\description{
More Unique IDs Derived from sampleData
}
\keyword{datasets}
