% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Data from NLSY}
\format{
A data frame. \describe{Kinship pairs and their relatedness, height,
  and weight information.}
}
\source{
NLSY/R Lab
}
\usage{
sample_data
}
\description{
A data frame output from \link[NlsyLinks]{CreatePairLinksSingleEntered} that
contains data for kinship pairs' height and weight.
}
\keyword{datasets}
