% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data2.R
\docType{data}
\name{sample_data2}
\alias{sample_data2}
\title{Sample Data for analysis}
\format{
A data frame. \describe{pairs and their height,
  weight, marriage status, test results, and dependent variable (DV) information.}
}
\usage{
sample_data2
}
\description{
A data frame with 11 columns
contains dyadic data for pair's dependent variable (DV), height, weight, marriage status, and test results.
}
\keyword{datasets}
