% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleData}
\alias{sampleData}
\title{Sample Data from NLSY}
\format{
A data frame with 50 rows and 63 columns \describe{Sibling Pairs and
  their percentile for education, occupation, total net family income, and
  SES for the years 2002, 2004, 2006, and 2008. It also contains individuals'
  race and sex. For race, the NLSY categorizes are: }
}
\source{
NLSY/R Lab
}
\usage{
sampleData
}
\description{
A data frame output from the original
\code{CreatePairLinksDoubleEntered} function that the original
\code{\link{discord_data}} function will work on. It contains percentile
information for education, occupation, total net family income, and a raw SES
score.
}
\keyword{datasets}
