% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniqueExtendedIDs}
\alias{uniqueExtendedIDs}
\title{Unique IDs Derived from sampleData}
\format{
A dataframe for testing with 32 rows
}
\source{
NLSY/R Lab
}
\usage{
uniqueExtendedIDs
}
\description{
A dataframe output from the original
\code{CreatePairLinksDoubleEntered} function that the original
\code{\link{discord_data}} function will work on. It contains percentile
information for education, occupation, total net family income, and a raw SES
score.
}
\keyword{datasets}
