% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_regression.R
\name{check_sibling_order}
\alias{check_sibling_order}
\title{Check which sibling has more of the outcome}
\usage{
check_sibling_order(data, outcome, pair_identifiers, row)
}
\arguments{
\item{data}{The data set with kinship pairs.}

\item{outcome}{A character string containing the outcome variable of
interest.}

\item{pair_identifiers}{A character vector of length two that contains the
variable identifier for each kinship pair.}

\item{row}{The row number of the data frame}
}
\value{
A character string signifying which familial member (1, 2, or
  neither) has more of the outcome.
}
\description{
This function dds a column \code{order} by comparing which familial member
has more of the outcome. This is done per pair (i.e. row).
}
