% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord-package.R
\docType{package}
\name{discord-package}
\alias{discord}
\alias{discord-package}
\title{discord: Functions for Discordant Kinship Modeling}
\description{
Functions for discordant kinship modeling (and other sibling-based quasi-experimental designs). Contains data restructuring functions and functions for generating biometrically informed data for kin pairs. See [Garrison and Rodgers, 2021 <>], [Sims, Trattner, and Garrison, 2024 <>) for empirical examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/R-Computing-Lab/discord}
  \item \url{https://r-computing-lab.github.io/discord/}
}

}
\author{
\strong{Maintainer}: S. Mason Garrison \email{garrissm@wfu.edu} (\href{https://orcid.org/0000-0002-4804-6003}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jonathan Trattner \email{code@jdtrat.com} (\href{https://orcid.org/0000-0002-1097-7603}{ORCID}) (https://www.jdtrat.com/)
  \item Hwang Yoo Ri \email{yrhwang89@gmail.com}
}

Other contributors:
\itemize{
  \item Cermet Ream [contributor]
}

}
\keyword{internal}
