% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord_data.R
\name{discord_data}
\alias{discord_data}
\title{Restructure Data}
\usage{
discord_data(doublentered = F, outcome = y, sep = "", scale = T, df,
  id = NULL)
}
\arguments{
\item{outcome}{name of your outcome variable}

\item{scale}{rescale all variables at the individual level to have a mean of 0 and a SD of 1}

\item{df}{dataframe with all variables in it}

\item{id}{optional id variable}

\item{doubleentered}{is data double entered? Default is FALSE}
}
\value{
Returns \code{data.frame} with the following:
\item{id}{id}
\item{outcome_1}{outcome for kin1; kin1 is always greater than kin2, except when tied. Then kin1 is randomly selected from the pair}
\item{outcome_2}{outcome for kin2}
\item{outcome_diff}{difference between outcome of kin1 and kin2}
\item{outcome_mean}{mean outcome for kin1 and kin2}
\item{predictor_i_1}{predictor variable i for kin1}
\item{predictor_i_2}{predictor variable i for kin2}
\item{predictor_i_diff}{difference between predictor i of kin1 and kin2}
\item{predictor_i_mean}{mean predictor i for kin1 and kin2}
}
\description{
Restructure wide form data into analyzable data, sorted by outcome.
}

