% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord_regression.R
\name{discord_regression}
\alias{discord_regression}
\title{Discord Regression}
\usage{
discord_regression(
  df,
  outcome,
  predictors,
  more_args = NULL,
  additional_formula = more_args,
  ...
)
}
\arguments{
\item{df}{dataframe with all variables in it.}

\item{outcome}{Name of outcome variable}

\item{predictors}{Names of predictors.}

\item{more_args}{Optional string to add additional inputs to formula}

\item{additional_formula}{Depreciated}

\item{...}{Optional pass on additional inputs.}
}
\value{
Returns \code{lm}
}
\description{
Function to apply discordant regression to discordant data.
}
