% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord_regression.R
\name{discord_regression}
\alias{discord_regression}
\title{Discord Regression}
\usage{
discord_regression(doublentered = F, outcome = y, sep = "", scale = T,
  df, id = NULL)
}
\arguments{
\item{r_all}{levels of relatedness; default is MZ and DZ twins c(1,.5)}
}
\value{
Returns \code{data.frame} with the following:
\item{Ai_1}{genetic component for variable i for kin1}
}
\description{
Run discord analysis on discord data
}

