% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord_regression.R
\name{discord_regression}
\alias{discord_regression}
\title{Discord Regression}
\usage{
discord_regression(doublentered = F, outcome = y, sep = "", scale = T,
  df = NULL, id = NULL)
}
\arguments{
\item{x}{x}
}
\value{
Returns \code{data.frame} with the following:
\item{X}{X}
}
\description{
Run discord analysis on discord data.
}

