% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_discord_regression.R
\name{discord_regression}
\alias{discord_regression}
\title{Perform a Linear Regression within the Discordant Kinship Framework}
\usage{
discord_regression(
  data,
  outcome,
  predictors,
  id = "extended_id",
  sex = "sex",
  race = "race",
  pair_identifiers = c("_s1", "_s2"),
  abridged_output = FALSE,
  legacy = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{outcome}{A character string containing the outcome variable of
interest.}

\item{predictors}{A character vector containing the column names for
predicting the outcome.}

\item{id}{A unique kinship pair identifier.}

\item{sex}{A character string for the sex column name.}

\item{race}{A character string for the race column name.}

\item{pair_identifiers}{A character vector of length two that contains the variable identifier for each kinship pair.}

\item{abridged_output}{Logical: FALSE (by default) and the fit model will be summarized with the \link[broom]{tidy} function. FALSE and the full model object will be returned.}

\item{legacy}{Logical Logical: FALSE (by default) when true uses legacy code version}
}
\value{
Either a tidy data frame containing the model metrics or the full model object will be returned. See examples.
}
\description{
Perform a Linear Regression within the Discordant Kinship Framework
}
\examples{

# Return an abridged model output using the \link[broom]{package}.
discord_regression(data = sample_data,
outcome = "height",
predictors = "weight",
pair_identifiers = c("_s1", "_s2"),
sex = NULL,
race = NULL)

# Return the full model output.
discord_regression(data = sample_data,
outcome = "height",
predictors = "weight",
pair_identifiers = c("_s1", "_s2"),
sex = NULL,
race = NULL,
abridged_output = FALSE)

}
