% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discord_data_JT.R
\name{discordData}
\alias{discordData}
\title{Prepare data for regression analysis}
\usage{
discordData(data, variables, id, sex, race)
}
\arguments{
\item{data}{The output of \link[Nlsylinks]{CreatePairLinksDoubleEntered}.}

\item{variables}{A character vector of variables to transform.}

\item{id}{The extended family pair ID from
\link[Nlsylinks]{CreatePairLinksDoubleEntered}. Should be a character
vector.}

\item{sex}{The character string for the sex column name.}

\item{race}{The character string for the race column name.}
}
\value{
A data frame that contains one row for each family pair and the difference
and mean of their scores for different variables.
}
\description{
This function takes in the data and transforms it into a format where a
discord regression can be performed.
}
