% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_regression.R
\name{make_mean_diffs}
\alias{make_mean_diffs}
\title{Make Mean Differences}
\usage{
make_mean_diffs(..., fast = FALSE)
}
\arguments{
\item{...}{Additional arguments to be passed to the function.}

\item{fast}{Logical. If TRUE, uses a faster method for data processing.}
}
\description{
This function calculates differences and means of a given variable for each kinship pair. The order of subtraction and the variables' names in the output dataframe depend on the order column set by check_sibling_order().
If the demographics parameter is set to "race", "sex", or "both", it also prepares demographic information accordingly,
swapping the order of demographics as per the order column.
}
