% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_regression.R
\name{make_mean_diffs}
\alias{make_mean_diffs}
\title{Make Mean Differences}
\usage{
make_mean_diffs(
  data,
  id,
  sex,
  race,
  demographics,
  variable,
  pair_identifiers,
  row
)
}
\description{
This function calculates differences and means of a given variable for each kinship pair. The order of subtraction and the variables' names in the output dataframe depend on the order column set by check_sibling_order(). 
If the demographics parameter is set to "race", "sex", or "both", it also prepares demographic information accordingly, 
swapping the order of demographics as per the order column.
}
