% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_restructure.R
\name{restructure}
\alias{restructure}
\title{Restructure dyadic-structured data for discordant kinship modeling. This function restructures data to determine the mean score(s) and the difference score(s) of the pair. Except for the pair id, all the variable names should end with a pair identifier (either number 1 or number 2) to use this function.}
\usage{
restructure(
  data,
  pid,
  num_con = 1,
  continuous_var1_1,
  continuous_var1_2,
  continuous_var2_1 = NULL,
  continuous_var2_2 = NULL,
  num_cat = 1,
  categorical_var1_1,
  categorical_var1_2,
  categorical_var2_1 = NULL,
  categorical_var2_2 = NULL,
  y_1,
  y_2
)
}
\arguments{
\item{data}{A data frame.}

\item{pid}{A numeric vector indicating unique pair identifiers.}

\item{num_con}{The number of continuous variable(s) (independent variable); default is 1. At present, it is limited to max of two variables.}

\item{continuous_var1_1}{A numeric vector containing the independent variable of the member 1.}

\item{continuous_var1_2}{A numeric vector containing the independent variable of the member 2.}

\item{continuous_var2_1}{A numeric vector containing the other independent variable of the member 1; default is NULL.}

\item{continuous_var2_2}{A numeric vector containing the other independent variable of the member 2; default is NULL.}

\item{num_cat}{The number of categorical variable(s) (independent variable); default is 1. At present, it is limited to max of two variables.}

\item{categorical_var1_1}{A categorical vector containing the independent variable of the member 1.}

\item{categorical_var1_2}{A categorical vector containing the independent variable of the member 2.}

\item{categorical_var2_1}{A categorical vector containing the other independent variable of the member 1; default is NULL.}

\item{categorical_var2_2}{A categorical vector containing the other independent variable of the member 2; default is NULL.}

\item{y_1}{The dependent variable of the member 1.}

\item{y_2}{The dependent variable of the member 2.}
}
\value{
A data frame that contains analyzable, dyadic-structured data for performing
  kinship regressions.
}
\description{
Restructure dyadic-structured data for discordant kinship modeling. This function restructures data to determine the mean score(s) and the difference score(s) of the pair. Except for the pair id, all the variable names should end with a pair identifier (either number 1 or number 2) to use this function.
}
\examples{

restructure(data = sample_data2,
pid = 'pairID',
num_con = 1,
continuous_var1_1 = 'height_1',
continuous_var1_2 = 'height_2',
continuous_var2_1 = NULL,
continuous_var2_2 = NULL,
num_cat = 1,
categorical_var1_1 = 'marriage_1',
categorical_var1_2 = 'marriage_2',
categorical_var2_1 = NULL,
categorical_var2_2 = NULL,
y_1 = 'DV_1',
y_2 = 'DV_2')
}
